requirements.txt
pyrebase4==4.5.0
gaeconf==0.1.3
Pillow==9.5.0
python_jwt
gcloud 
sseclient
pycryptodome
requests-toolbelt==0.8.0
requests==2.22.0
urllib3==1.25.11
feedparser==6.0.2
Flask==3.0.0